requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
auto-py-to-exe==2.44.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0.4
docstring_parser==0.16
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.0.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json-streamer==0.1.0
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
openai-streaming==0.4.0
openpyxl==3.1.3
outcome==1.3.0.post0
overrides==7.7.0
package_name==0.1
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proxy_tools==0.1.0
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pythonnet==3.0.3
pytz==2024.1
pywebview==5.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
selenium==4.21.0
Send2Trash==1.8.2
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
trio==0.25.1
trio-websocket==0.11.1
twilio==9.1.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wfastcgi==3.0.0
whichcraft==0.6.1
widgetsnbextension==4.0.10
worksheet==0.0.1a0
wsproto==1.2.0
xdg==6.0.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
